src/main/java/com/sazaxa/shipmentapi/point/service/PointService.java
package com.sazaxa.shipmentapi.point.service;

import com.sazaxa.shipmentapi.point.dto.PointConfigRequestDto;
import com.sazaxa.shipmentapi.point.dto.PointConfigResponseDto;
import com.sazaxa.shipmentapi.point.repository.PointRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Transactional
@Service
public class PointService {
    private final PointRepository pointRepository;

    public PointService(PointRepository pointRepository) {
        this.pointRepository = pointRepository;
    }

    public PointConfigResponseDto configurePoint(PointConfigRequestDto pointConfigRequestDto) {
        return null;
    }
}
